requirements.txt
flask==3.1.1
flask-cors==5.0.1
openai==1.78.1
python-dotenv==1.1.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.169.0
firebase-admin==6.8.0
oauth2client==3.0.0
gunicorn==21.2.0 